src/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
streamlit
xgboost
scikit-learn
pandas
numpy
joblib
requests
